requirements.txt
asyncio
colorama
discord
discord-components
discord-py-slash-command
discord.py>=1.6.0
DiscordUtils
Pillow
PyAutoGUI
PyYAML=5.4.1
requests
requests-toolbelt
rich